Refactor DynatraceTag code into own file
