+++ test/extensions/tracers/opentelemetry/samplers/dynatrace/BUILD
@@ -34,6 +34,7 @@ envoy_extension_cc_test(
         "sampling_controller_test.cc",
         "stream_summary_test.cc",
         "tenant_id_test.cc",
+        "dynatrace_tag_test.cc",
     ],
     extension_names = ["envoy.tracers.opentelemetry.samplers.dynatrace"],
     rbe_pool = "6gig",
